Convert App to TS file
